[Feature #159847558] Allow user to vote and down vote an answer
