+++ app/answers/api/v1/view.py
@@ -83,42 +83,14 @@ class ListAPIView(MethodView):
         return (jsonify(response_object)), 200
 
 
-class VoteAPIView(MethodView):
-    """ Update Instance api resource """
-
-    @jwt_required
-    def post(self, answer_id=None):
-        post_data = request.get_json(force=True)
-        response = Table.vote(str(answer_id), data=post_data)
-        if response:
-            response_object = {
-                'status': 'success',
-                'message': 'Your vote was successful'
-            }
-            return make_response(jsonify(response_object)), 201
-
-        response_object = {
-            'status': 'fail',
-            'message': 'Some error occurred. Please try again.'
-        }
-        return make_response(jsonify(response_object)), 400
-
-
 # Define the API resources
 create_view = CreateAPIView.as_view('create_api')
 list_view = ListAPIView.as_view('list_api')
-vote_view = VoteAPIView.as_view('vote_api')
 
 # Add Rules for API Endpoints
 answers_blueprint.add_url_rule(
     '/api/v1/questions/<int:question_id>/answers',
     view_func=create_view,
-    methods=['POST']
-)
-
-answers_blueprint.add_url_rule(
-    '/api/v1/questions/answers/vote/<string:answer_id>',
-    view_func=vote_view,
     methods=['POST']
 )
