+++ app/votes/views.py
@@ -1,0 +1,3 @@
+#################
+# imports
+#################
